Refactor: Rename deploy_to_cv role and module
