+++ .github/release.yml
@@ -7,8 +7,8 @@ changelog:
       - 'rn: CI(cvp_configlet_upload)'
       - 'rn: Test(dhcp_provisioner)'
       - 'rn: CI(dhcp_provisioner)'
-      - 'rn: Test(deploy_to_cv)'
-      - 'rn: CI(deploy_to_cv)'
+      - 'rn: Test(cv_deploy)'
+      - 'rn: CI(cv_deploy)'
       - 'rn: Test(eos_cli_config_gen)'
       - 'rn: CI(eos_cli_config_gen)'
       - 'rn: Test(eos_config_deploy_cvp)'
@@ -41,7 +41,7 @@ changelog:
         - 'rn: Feat(build_output_folders)!'
         - 'rn: Feat(cvp_configlet_upload)!'
         - 'rn: Feat(dhcp_provisioner)!'
-        - 'rn: Feat(deploy_to_cv)!'
+        - 'rn: Feat(cv_deploy)!'
         - 'rn: Feat(eos_cli_config_gen)!'
         - 'rn: Feat(eos_config_deploy_cvp)!'
         - 'rn: Feat(eos_config_deploy_eapi)!'
@@ -56,7 +56,7 @@ changelog:
         - 'rn: Fix(build_output_folders)!'
         - 'rn: Fix(cvp_configlet_upload)!'
         - 'rn: Fix(dhcp_provisioner)!'
-        - 'rn: Fix(deploy_to_cv)!'
+        - 'rn: Fix(cv_deploy)!'
         - 'rn: Fix(eos_cli_config_gen)!'
         - 'rn: Fix(eos_config_deploy_cvp)!'
         - 'rn: Fix(eos_config_deploy_eapi)!'
@@ -71,7 +71,7 @@ changelog:
         - 'rn: Cut(build_output_folders)!'
         - 'rn: Cut(cvp_configlet_upload)!'
         - 'rn: Cut(dhcp_provisioner)!'
-        - 'rn: Cut(deploy_to_cv)!'
+        - 'rn: Cut(cv_deploy)!'
         - 'rn: Cut(eos_cli_config_gen)!'
         - 'rn: Cut(eos_config_deploy_cvp)!'
         - 'rn: Cut(eos_config_deploy_eapi)!'
@@ -86,7 +86,7 @@ changelog:
         - 'rn: Revert(build_output_folders)!'
         - 'rn: Revert(cvp_configlet_upload)!'
         - 'rn: Revert(dhcp_provisioner)!'
-        - 'rn: Revert(deploy_to_cv)!'
+        - 'rn: Revert(cv_deploy)!'
         - 'rn: Revert(eos_cli_config_gen)!'
         - 'rn: Revert(eos_config_deploy_cvp)!'
         - 'rn: Revert(eos_config_deploy_eapi)!'
@@ -101,7 +101,7 @@ changelog:
         - 'rn: Refactor(build_output_folders)!'
         - 'rn: Refactor(cvp_configlet_upload)!'
         - 'rn: Refactor(dhcp_provisioner)!'
-        - 'rn: Refactor(deploy_to_cv)!'
+        - 'rn: Refactor(cv_deploy)!'
         - 'rn: Refactor(eos_cli_config_gen)!'
         - 'rn: Refactor(eos_config_deploy_cvp)!'
         - 'rn: Refactor(eos_config_deploy_eapi)!'
@@ -116,7 +116,7 @@ changelog:
         - 'rn: Bump(build_output_folders)!'
         - 'rn: Bump(cvp_configlet_upload)!'
         - 'rn: Bump(dhcp_provisioner)!'
-        - 'rn: Bump(deploy_to_cv)!'
+        - 'rn: Bump(cv_deploy)!'
         - 'rn: Bump(eos_cli_config_gen)!'
         - 'rn: Bump(eos_config_deploy_cvp)!'
         - 'rn: Bump(eos_config_deploy_eapi)!'
@@ -149,7 +149,7 @@ changelog:
         - 'rn: Fix(build_output_folders)'
         - 'rn: Fix(cvp_configlet_upload)'
         - 'rn: Fix(dhcp_provisioner)'
-        - 'rn: Fix(deploy_to_cv)'
+        - 'rn: Fix(cv_deploy)'
         - 'rn: Fix(eos_config_deploy_cvp)'
         - 'rn: Fix(eos_config_deploy_eapi)'
         - 'rn: Fix(eos_snapshot)'
@@ -163,7 +163,7 @@ changelog:
         - 'rn: Doc(build_output_folders)'
         - 'rn: Doc(cvp_configlet_upload)'
         - 'rn: Doc(dhcp_provisioner)'
-        - 'rn: Doc(deploy_to_cv)'
+        - 'rn: Doc(cv_deploy)'
         - 'rn: Doc(eos_cli_config_gen)'
         - 'rn: Doc(eos_config_deploy_cvp)'
         - 'rn: Doc(eos_config_deploy_eapi)'
@@ -191,7 +191,7 @@ changelog:
         - 'rn: Feat(build_output_folders)'
         - 'rn: Feat(cvp_configlet_upload)'
         - 'rn: Feat(dhcp_provisioner)'
-        - 'rn: Feat(deploy_to_cv)'
+        - 'rn: Feat(cv_deploy)'
         - 'rn: Feat(eos_config_deploy_cvp)'
         - 'rn: Feat(eos_config_deploy_eapi)'
         - 'rn: Feat(eos_snapshot)'
