+++ ansible_collections/arista/avd/docs/_media/excalidraw/deploy_to_cv.excalidraw
@@ -31,11 +31,11 @@
       "locked": false,
       "fontSize": 20,
       "fontFamily": 3,
-      "text": "deploy_to_cv",
+      "text": "cv_deploy",
       "textAlign": "center",
       "verticalAlign": "top",
       "containerId": null,
-      "originalText": "deploy_to_cv",
+      "originalText": "cv_deploy",
       "lineHeight": 1.2,
       "baseline": 19
     },
@@ -444,11 +444,11 @@
       "locked": false,
       "fontSize": 16,
       "fontFamily": 3,
-      "text": "Modules:\n- arista.avd.deploy_to_cv\n",
+      "text": "Modules:\n- arista.avd.cv_deploy\n",
       "textAlign": "left",
       "verticalAlign": "top",
       "containerId": null,
-      "originalText": "Modules:\n- arista.avd.deploy_to_cv\n",
+      "originalText": "Modules:\n- arista.avd.cv_deploy\n",
       "lineHeight": 1.1875,
       "baseline": 53
     },
