+++ ansible_collections/arista/avd/plugins/modules/cv_workflow.py
@@ -2,19 +2,14 @@
 # Use of this source code is governed by the Apache License 2.0
 # that can be found in the LICENSE file.
 
-# NOTE: This is supposed to be deprecated as per
-# https://docs.ansible.com/ansible/latest/dev_guide/developing_modules_documenting.html#ansible-metadata-block
-# But our doc Jinja2 template renders it as preview which is what we want
-ANSIBLE_METADATA = {"metadata_version": "1.0", "status": ["preview"]}
-
 DOCUMENTATION = r"""
 ---
-module: deploy_to_cv
-version_added: "4.6.0"
+module: cv_workflow
+version_added: "4.7.0"
 author: Arista Ansible Team (@aristanetworks)
 short_description: Deploy various objects to CloudVision
 description: |-
-  The `arista.avd.deploy_to_cv` module is an Ansible Action Plugin providing the following capabilities:
+  The `arista.avd.cv_workflow` module is an Ansible Action Plugin providing the following capabilities:
 
   - Verify Devices are in the CloudVision inventory.
   - Verify Devices are in the Inventory & Topology Studio.
@@ -151,7 +146,7 @@ EXAMPLES = r"""
     - name: Provision CVP with AVD configuration
       run_once: true
       delegate_to: localhost
-      arista.avd.deploy_to_cv:
+      arista.avd.cv_workflow:
         cv_servers: [ "www.arista.io" ]
         cv_token: "<insert vaulted service account token here>"
         # cv_verify_certs: True
