+++ ansible_collections/arista/avd/roles/cv_deploy/defaults/main/directories.yml
