+++ ansible_collections/arista/avd/roles/cv_deploy/defaults/main/main.yml
@@ -53,7 +53,7 @@ cv_strict_tags: false
 # Set the template to be used to generate the configlet names in CloudVision Static Config Studio.
 cv_configlet_name_template: "AVD-${hostname}"
 
-# If true, detailed deployment results will be registered into 'deploy_to_cv_results' variable.
+# If true, detailed deployment results will be registered into 'cv_deploy_results' variable.
 # Otherwise only the basic result like 'failed', 'warnings' and 'errors' are registered.
 # There is a small performance impact on this, which is why it is not registered by default.
 cv_register_detailed_results: false
