+++ ansible_collections/arista/avd/roles/cv_deploy/meta/main.yml
@@ -1,0 +1,3 @@
+galaxy_info:
+  description: Deploy device configurations and tags to CloudVision
+  standalone: false
