+++ ansible_collections/arista/avd/roles/deploy_to_cv/meta/main.yml
@@ -1,3 +1,3 @@
 galaxy_info:
-  description: Deploy device configurations and tags to CloudVision
+  description: DEPRECATED - Migrated to arista.avd.cv_deploy
   standalone: false
