+++ ansible_collections/arista/avd/tests/sanity/ignore-2.14.txt
@@ -10,4 +10,4 @@ plugins/modules/yaml_templates_to_facts.py validate-modules:missing-gplv3-licens
 plugins/vars/global_vars.py validate-modules:missing-gplv3-license
 plugins/modules/eos_validate_state_runner.py validate-modules:missing-gplv3-license
 plugins/modules/eos_validate_state_reports.py validate-modules:missing-gplv3-license
-plugins/modules/deploy_to_cv.py validate-modules:missing-gplv3-license
+plugins/modules/cv_deploy.py validate-modules:missing-gplv3-license
